+++ packages/types/src/client-types.ts
@@ -18,6 +18,15 @@ export interface IRequestData extends Omit<RequestLogic.IRequest, 'currency'> {
   pending: RequestLogic.IPendingRequest | null;
 }
 
+/** Interface request data with event emitter and subscriber */
+export interface IRequestDataWithEvents extends IRequestData {
+  on: <K extends keyof IRequestEvents>(event: K, listener: IRequestEvents[K]) => this;
+  emit: <K extends keyof IRequestEvents>(
+    event: K,
+    ...args: Parameters<IRequestEvents[K]>
+  ) => boolean;
+}
+
 /** Create request parameters */
 export interface ICreateRequestParameters {
   requestInfo: RequestLogic.ICreateParameters | IRequestInfo;
@@ -36,4 +45,9 @@ export interface IRequestInfo {
   extensionsData?: any[];
   timestamp?: number;
   nonce?: number;
+}
+
+/** Events types risen by a request */
+export interface IRequestEvents {
+  confirmed: (requestData: IRequestDataWithEvents) => void;
 }